images-api/requirements.txt
Flask==2.3.2
gunicorn==21.2.0
google-cloud-storage==2.9.0
